chore(LocationService): remove Google-related code
