+++ config/dotcom/dotcom.exs
@@ -34,8 +34,6 @@ config :dotcom, route_populate_caches?: config_env() == :prod
 routes_repo = if config_env() == :test, do: Routes.MockRepoApi, else: Routes.Repo
 config :dotcom, :routes_repo_api, routes_repo
 
-config :dotcom, :location_http_pool, :google_http_pool
-
 repo_module = if config_env() == :test, do: RoutePatterns.MockRepo, else: RoutePatterns.Repo
 
 config :dotcom, :route_patterns_repo_api, repo_module
