+++ config/runtime.exs
@@ -192,9 +192,6 @@ config :dotcom, LocationService,
   google_api_key: System.get_env("GOOGLE_API_KEY"),
   google_client_id: System.get_env("GOOGLE_MAPS_CLIENT_ID") || "",
   google_signing_key: System.get_env("GOOGLE_MAPS_SIGNING_KEY") || "",
-  geocode: System.get_env("LOCATION_SERVICE") || "aws",
-  reverse_geocode: System.get_env("LOCATION_SERVICE") || "aws",
-  autocomplete: System.get_env("LOCATION_SERVICE") || "aws",
   aws_index_prefix: System.get_env("AWS_PLACE_INDEX_PREFIX", "dotcom-prod")
 
 config :dotcom, DotcomWeb.ViewHelpers,
