+++ lib/dotcom_web/controllers/schedule/map_api.ex
@@ -33,7 +33,7 @@ defmodule DotcomWeb.ScheduleController.MapApi do
   @spec get_map_data(Route.t()) :: MapData.t()
   defp get_map_data(route) do
     route_patterns = LineHelpers.get_map_route_patterns(route.id, route.type)
-    {_map_img_src, dynamic_map_data} = Maps.map_data(route, [], [], route_patterns, [])
+    {_map_img_src, dynamic_map_data} = Maps.map_data(route, route_patterns, [])
     dynamic_map_data
   end
 end
