+++ lib/dotcom_web/router.ex
@@ -253,7 +253,7 @@ defmodule DotcomWeb.Router do
   scope "/places", DotcomWeb do
     pipe_through([:api])
 
-    get("/autocomplete/:input/:hit_limit/:token", PlacesController, :autocomplete)
+    get("/autocomplete/:input/:hit_limit", PlacesController, :autocomplete)
     get("/details/:address", PlacesController, :details)
     get("/reverse-geocode/:latitude/:longitude", PlacesController, :reverse_geocode)
