+++ lib/google_maps/map_data.ex
