+++ lib/google_maps/map_data/marker.ex
