+++ lib/google_maps/map_data/path.ex
