+++ lib/google_maps/map_data/symbol.ex
