+++ lib/google_maps/view_helpers.ex
