--- lib/location_service/google_maps/geocode/input.ex
@@ -1,8 +1,0 @@
-defmodule GoogleMaps.Geocode.Input do
-  @type t :: %__MODULE__{
-          address: String.t() | nil,
-          latitude: float | nil,
-          longitude: float | nil
-        }
-  defstruct [:address, :latitude, :longitude]
-end
