+++ lib/location_service/suggestion.ex
@@ -1,7 +1,6 @@
 defmodule LocationService.Suggestion do
   @moduledoc """
   An autocomplete suggestion.
-  [todo] deprecate GoogleMaps.Place.Prediction
   """
   @type t :: %__MODULE__{
           address: String.t(),
