+++ test/location_service/result_test.exs
@@ -4,77 +4,42 @@ defmodule LocationService.ResultTest do
   use ExUnit.Case, async: false
   import ExUnit.CaptureLog
   import LocationService.Result
+  import Test.Support.EnvHelpers
 
   @input "anything really"
 
   setup do
-    aws_decoded_body = %{
-      "Results" => [
-        %{
-          "Place" => %{
-            "Label" => "Somewhere",
-            "Geometry" => %{"Point" => [-71.05566, 42.35913]}
+    {:ok, encoded_body} =
+      Jason.encode(%{
+        "Results" => [
+          %{
+            "Place" => %{
+              "Label" => "Somewhere",
+              "Geometry" => %{"Point" => [-71.05566, 42.35913]}
+            }
           }
-        }
-      ]
-    }
+        ]
+      })
 
-    {:ok, aws_encoded_body} = Jason.encode(aws_decoded_body)
-    {:ok, google_decoded_body} = '{
-      "address_components" : [{
-          "long_name" : "52-2",
-          "short_name" : "52-2",
-          "types" : [ "street_number" ]
-        }],
-      "formatted_address" : "52-2 Park Ln, Boston, MA 02210, USA",
-      "geometry" : {
-        "bounds" : {
-          "northeast" : {
-            "lat" : 42.3484946,
-            "lng" : -71.0389612
-          },
-          "southwest" : {
-            "lat" : 42.3483114,
-            "lng" : -71.03938769999999
-          }
-        },
-        "location" : {
-          "lat" : 42.3484012,
-          "lng" : -71.039176
-        }
-      }
-    }' |> Jason.decode()
+    {:ok, empty_encoded_body} = Jason.encode(%{"status" => "ZERO_RESULTS"})
 
-    old_level = Logger.level()
-
-    on_exit(fn ->
-      Logger.configure(level: old_level)
-    end)
-
-    Logger.configure(level: :info)
+    set_log_level(:info)
 
     %{
-      google_decoded_body: google_decoded_body,
-      aws_encoded_body: aws_encoded_body
+      encoded_body: encoded_body,
+      empty_encoded_body: empty_encoded_body
     }
   end
 
   describe "handle_response/2" do
     test "handles and parses a result with a body and 200 status code", %{
-      google_decoded_body: google_decoded_body,
-      aws_encoded_body: aws_encoded_body
+      encoded_body: encoded_body
     } do
       log =
         capture_log(fn ->
-          aws_result =
-            {:ok, %{status_code: 200, body: aws_encoded_body}} |> handle_response(@input)
-
-          google_result =
-            {:ok, %{"status" => "OK", "results" => [google_decoded_body]}}
-            |> handle_response(@input)
+          aws_result = {:ok, %{status_code: 200, body: encoded_body}} |> handle_response(@input)
 
           assert {:ok, [%LocationService.Address{} | _]} = aws_result
-          assert {:ok, [%LocationService.Address{} | _]} = google_result
         end)
 
       assert log =~ "[info]"
@@ -82,11 +47,11 @@ defmodule LocationService.ResultTest do
     end
 
     test "returns an error for a result with a non-200 code", %{
-      aws_encoded_body: aws_encoded_body
+      encoded_body: encoded_body
     } do
       log =
         capture_log(fn ->
-          result = {:ok, %{status_code: 416, body: aws_encoded_body}} |> handle_response(@input)
+          result = {:ok, %{status_code: 416, body: encoded_body}} |> handle_response(@input)
 
           assert {:error, :internal_error} = result
         end)
@@ -99,10 +64,11 @@ defmodule LocationService.ResultTest do
     test "returns an error for no results" do
       log =
         capture_log(fn ->
-          assert {:error, :zero_results} = {:ok, %{"Results" => []}} |> handle_response(@input)
+          assert {:error, :zero_results} =
+                   {:ok, %{"status" => "ZERO_RESULTS"}} |> handle_response(@input)
 
           assert {:error, :zero_results} =
-                   {:ok, %{"status" => "OK", "results" => []}} |> handle_response(@input)
+                   {:ok, %{"status" => "OK", "Results" => []}} |> handle_response(@input)
         end)
 
       assert log =~ "[info]"
